# swagger/jmarkets.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JMarkets API Documentation",
    "version": "1.0.0",
    "description": "Multi-tenant marketplace platform API with CMS, authentication, RBAC, and auto-deployment",
    "contact": {
      "name": "JMarkets Support",
      "url": "https://j-markets.jcampos.dev"
    }
  },
  "servers": [
    {
      "url": "https://markets-api.jcampos.dev",
      "description": "Production API"
    },
    {
      "url": "http://localhost:5000",
      "description": "Development API"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "paths": {
...
    "/api/users/{userId}/organization/{orgId}/upload/presigned": {
      "post": {
        "summary": "Get presigned URL for S3 upload",
        "tags": [
          "Upload"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fileName",
                  "fileType"
                ],
                "properties": {
                  "fileName": {
                    "type": "string"
                  },
                  "fileType": {
                    "type": "string"
                  },
                  "folder": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Presigned URL generated"
          }
        }
      }
    },
...
  },
...
}